Create a backwards compatible MerlinLog, keeping the setter in MerlinBuilder.
